feat: replace old handler
